Fix inout deshadowing so that it kicks in on throwing functions.
If we don't deshadow, we end up with reference counts that are larger
than expected and uniqueness checks fail for COW types when they should
pass (and this results in extra copies to ensure uniqueness). Of course
we also end up with extra copies entering and exiting functions, which
is bad by itself.

Fixes rdar://problem/21371013.

Swift SVN r29931