Collect inherited associated type conformances.
When we're determining which protocols an associated type conforms to,
look for same-named associated types in inherited protocols and add
their conformances to the set of protocols to which this associated
type conforms. Coupled with an embarrassing fix to the
protocol-minimization code, this should eliminate the dreaded SILGen
assertion about the number of conformances not matching the number of
protocols when generating witness table entries for associated type
witnesses. Fixes rdar://problem/21599502.

Swift SVN r29976