When type checking an arbitrary subexpression of the current node, we can run into trouble when we dive into a subexpr of a ClosureExpr, because that subexpr may refer to type variables on the closureexpr's parameters.
Check for this case, and refuse to dive into the subexpr in this case.  It would
be great to rewrite the closure parameter types to Type() or ErrorType or something
so that we can proceed even in this case, but this causes us to fail to catch
nested constraint checking failures.

This was figured out while working on other things, but fixes a validation test.



Swift SVN r30135