Builtin.reinterpretCast should retain any references it generates.
This fixes a bug where reinterpret cast would cause an object to be
freed prematurely. In the common case of casting a known reference
type to another reference type, we still elide the extra retain/release.

And finally, this builtin should have identical semantics for concrete
types and unbounds generics.

Swift SVN r30217