[llvm-arc-opts] Merge together multiple retains/releases in the same BB into retain_n/release_n calls.
We are taking advantage of the fact that we can always safely move a retain
earlier in the program and a release later in the program. Given this
information, in every BB, we merge together all retains into the first retain
(creating a retain_n) and all releases into the last release in that BB
(creating a release_n).

rdar://21803771

Swift SVN r30261