Dependencies: Break out provided members from other ways to provide/extend a type.
There's now a difference between "this file provides a particular type" and
"this file provides members on a particular type". Adding a conformance in
an extension is still considered "providing a type" because it can affect
things besides just lookups.

This modifies the swiftdeps file output (including the members used by a file),
but the driver hasn't been updated to take advantage of this yet.

Swift SVN r30285