Sema: If a property type was inferred, copy it to the accessor before checking generic signature
Unlike trivial stored properties, observed properties get their accessors
created early on for some reason, before we have a type for the property.

If the type of the property was inferred and it was in a generic context,
we would then try to validate the generic signature before setting the
type of the setter's newValue parameter. Change the order to fix this.

Fixes <rdar://problem/21476759>.

Swift SVN r30381