SILGen: Fix crash with @objc override of non-@objc property
For properties, "must be accessed with ObjC dispatch" is different from
"needs objc metadata". When we mean the latter, just call isObjC() on
the AbstractStorageDecl instead of using hasObjCMethodDispatch().

The latter returns false if any overridden declaration is not @objc, in
which case at runtime we don't need to dispatch.

Fixes <rdar://problem/21544588>.

Swift SVN r30382