Fix for assert trying to devirtualize a try_apply during inlining.
When calling the devirtualizer from the inliner we were assuming that we
would only see ApplyInst and not TryApplyInst. That's not the case, and
since the devirtualizer has not been updated to handle
TryApplyInst (rdar://problem/21909405), we need to bail out if we see a
FullApplySite that is not a plain ApplyInst.

I hit this in our build with some local changes that I have.

Fixes rdar://problem/21913097.

Swift SVN r30441