Introduce a simple dedicated SIL pass to fetch and use pre-specializations at -Onone.
The logic which fetches pre-specialisations at -Onone is moved from the GenericSpecializer into a dedicated SIL pass called UsePrespecialized. This new pass is much simpler than GenericSpecializer. There are no changes to the overall logic.

The goal of this re-factoring is to pave the way for deletion of the stand-alone GenericSpecializer, because inline-devirt-specialize logic is now integrated thanks to Mark's recent improvements.

rdar://21881337

Swift SVN r30603